refactor oob teleporting to use the modular SO system
