+++ UOP1_Project/Assets/ScriptableObjects/StateMachine/Protagonist/Actions/TeleportIfFallingOutOfBounds.asset
@@ -1,0 +1,14 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 0d71147fadd9c3a42b47c399311e9e2c, type: 3}
+  m_Name: TeleportIfFallingOutOfBounds
+  m_EditorClassIdentifier:
