+++ UOP1_Project/Assets/ScriptableObjects/StateMachine/Protagonist/Actions/TeleportIfFallingOutOfBounds.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8b57d0c8d3b77e94d8ee4f92d492c39b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
