+++ UOP1_Project/Assets/Scripts/Characters/StateMachine/Actions/TeleportIfFallingOutOfBoundsSO.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0d71147fadd9c3a42b47c399311e9e2c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
